training/requirements.txt
torch==2.1.0
torchvision==0.16.0
onnx==1.15.0
onnxruntime==1.16.0
numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
scikit-learn==1.3.0
matplotlib==3.7.2
tqdm==4.66.1
